fix: bug in DoeSampleGenerator
The get_options_and_default_values method was filtering out the required
options from the algo_options_default dict. But this results in these
options not being added to the dynamic inputs => bug (e.g. n_samples for
lhs algo, which would not be passed from the config_dict to the
soswrapper).