fix: bug in disc.linearize (wip)
The new linearize method in GEMSEO bypasses the overridden get_input_data_names of SoSDiscipline,
and tries to compute the jacobian for all inputs, even non-numeric ones.
I overrid the add_differentiated_inputs/outputs of SoSDiscipline and SoSMDAChain,
but test 80/4 still fails on check_jacobian because the analytical and approx jacobians do not have the same key.