test: fix l0_001
Test 09 would fail because there was a difference of 10e-16 between the cached and input value of a coupling variable in study_2.
Since the cache tolerance is 0.0, this would trigger a new evaluation of the disciplines.
This is probably due to the new implementation of SoSScenarioAdapter: the input y_1 of study_2 is the final value of y_1 returned by study_1, while the cached value is the input value. Since the MDA tolerance is 1e-12, there is a small difference between these two values. By setting the MDA tolerance to 0.0, the input and output values of y_1 in study_1 are perfectly equal, so study_2 does not trigger any discipline execution anymore, since it's inputs are exactly the cached values.